Disable devicePixelRatio in PlasmaShell
Devive pixel ratio has some problems in plasmashell currently.

In particular:
- dialog continually expands (347951)
- Text element text is screwed (QTBUG-42606)
- Panel struts (350614)

It's also not a /huge/ deal as we do our own thing anyway, so the
experience is better to just turn off this scaling.

I've had this patch in my local code since forever.We should revert this
when all 3 are fixed.

CCBUG: 347951
REVIEW: 124463