[digital-clock] Add a simple option to use 24h-clock format
In my opinion this is just a temporary workaround and not a very good
one (generally, code-wise it's awesome :P). There should really be a
global config option for that but as we now rely on QLocale and QLocale
is...bad at this, here's at least this small help, which will at least
change the clock format on the panel.

Imo we could really really use something like KLocale again.

REVIEW: 124453
BUG: 345378
FIXED-IN: 5.4.0