Fix desktop context property for Desktop.qml
This is fallout from the shared engine changes.

With the shared view, engine()->rootContext() is not the same as
rootContext(). One being global, the other being per window.

This resulted us in us overwriting the same "desktop" context property,
and so Desktop.qml always had the property desktop pointing to the last
created desktop view.

The most obvious symptom is that the add widgets dialog appeared on the
wrong screen.

BUG: 351292
REVIEW: 124792