Fix lockup in digital clock
property tooSmall depends on height

If it's tool small it changes what text is shown on top of each other
Whether text is shown changes the height; super complex binding loop
that because it throws in a visibility change and goes via layouts isn't
being caught.

From what I can tell of other code when vertical we always want to show
both bits of text.

BUG: 351479
REVIEW: 124874