Monitor for the clock changes from the kernel
At the moment we update the clock when we resume from suspend or from
the time KCM.

However the signal from the clock KCM isn't very good as it signals just
after it requests an NTP update or clock change, not necessarily when
that finishes.

Also we miss any external NTP updates which might occur at other times
throughout the day, such as a big change when we connect to the network

This method also avoids needing solid to tell us when we're resuming
from suspend, which is deprecated API and also seems slightly slower.

BUG: 344870
REVIEW: 125028