check service from KOpenWithDialog is really valid
KOpenWithDialog when it finds an exec that matches looks for existing
.desktop files in this case it finds firefox, but detects that the exec
lines don't match.

In this case it creates a new service with a valid name copied from the
original .desktop file, but an invalid entryPath as it's not a direct
1-1 copy.

This code assume if we have a valid name, it must be a valid existing
.desktop file and uses the entryPath. By adding a check we explicitly
make a copy of the .desktop path.

BUG: 346909