[lockscreen] Improve autotests for lockwindow
Earlier we were showing dummy window to make sure that if background of
X is black and we fail to show our lockwindow, tests wouldn't assume
that lockwindow is shown.

However this solution had flow that window we were showing was not
colored at all. So either it was black or transperant and tests would
pass even if something is broken. This commit fixes it with :

- Show dummy window with red color
- Verify that red colored window is shown
- Show our lock window
- Verify that screen is black

REVIEW: 125805