Proxy Xembed icons to SNI
The goal of this project is to make xembed system trays available in
Plasma.

This is to allow legacy apps (xchat, pidgin, tuxguitar) etc. system
trays[1] available in Plasma which only supports StatusNotifierItem [2].

Ideally we also want this to work in an xwayland session, making X
system tray icons available even when plasmashell only has a wayland
connection.

REVIEW: 125655