[ksmserver] Add a dedicated --no-lockscreen command line option
On Wayland ksmserver should not be responsible for the lockscreen,
but the compositor takes care of providing the lock screen. To prevent
conflicts KSMServer gets a new command line option to enable it.

The code is slightly changed to not pass a series of boolean flags to
KSMServer, but a flags type.

The startplasma script for Wayland is adjusted to start ksmserver with
the new command line option.

REVIEW: 126022