Revert all of the Calendar-agenda changes
Because this code depends on plasma-framework and kdeclarative versions
that will not be released before Plasma 5.5 beta, this needs to be
reverted.

And because beta is also feature and string freeze, this will be
postponed to 5.6.

Reverts these commits:
a179d512b4fcba909a7c03da995dd963c220491b
69e13f67379e3f45e216a8f2423e56693674bf43
4b631f8c5c99957313ce2a1953b7975836fcef46
7232b798d557fb8d785827143aa576b24e790acc
ea6691a307f22490030a4fbfcc8d85535cc32284