Do not produce negative struts on switching screens
Turns out that this method is sometimes called at the time where the
virtualGeometry() is invalid -- which sounds pretty much like a Qt bug,
but it isn't exactly the first bug in screen handling in Qt5. Rather
than producing values which are surely invalid, try to cover up the
mess.

REVIEW: 126216