[notifications] Rework the notifications positioning a bit
The notifications popup positioning recently regressed
by some other changes (looks like Qt) and the popups
would fly across the screen.

The proper solution is using KWin effect but given how
close the release is, this needs to be dealt with in a
different way.

The main problem is calculating the initial popup size
because as long as the Dialog is invisible, it has an
incorrect geometry, so it needs to be positioned right
after it's being displayed. The Dialog however gets the
sizes even later, so the code now calls a slot from Dialog
that ensures the sizes are correct before the initial
placement on screen. It's not ideal but I'm out of ideas
otherwise. Plus it should be only temporary until the
KWin effect will replace it.

Besides that, it moves the custom position handling from the
QML into the C++ applet class, making it easier.

REVIEW: 126221
BUG: 355069
FIXED-IN: 5.5.0