Set root context properties on the right context
engine()->rootContext() will give a different result to
rootContext() when using a shared engine.

One being in the right context, the other being the context of the
shared engine. Using the latter means two panels end up sharing the same
object which leads to some interesting results and QML crashes in
QV4::QObjectWrapper::wrap when one gets deleted.

I've tested against 356545, but I think it's the cause of a lot of the
mystery QML bugs we couldn't figure out.

BUG: 356545
BUG: 355885
BUG: 356916
REVIEW: 126491