[DrKonqi] Be less intrusive
Instead of showing the full DrKonqi dialog, drawing a lot of attention to a crash,
a Status Notifier Item will be placed in system tray. For auto-restarting system
services like Plasma a crash would often go unnoticed if it weren't for the DrKonqi
dialog showing up. Additionally, regular application crashes will spawn a desktop
notification offering to report a bug or restart the application.

The Status Notifier Item hides after one minute because if the user didn't bother
within that time he probably forgot what he did to cause the crash anyway. If the
session is idle the SNI stays indefinitely so a crash of a long over-night task
like video rendering can still be reported.

If no notification service is running and the application did not auto-restart
regular DrKonqi will show up right away.

REVIEW: 126515