[Device Notifier] Rewrite delegates to use Layouts
The code is from a time before we had smart layouts and thus is pretty complicated
with lots of anchoring. This cleans up the code massively.

The layout is a bit tighter now, the eject/mount button is now a proper ToolButton.
The "n actions for this device" text now disappears when the actions are expanded;
(dis)appearance of the disk usage bar is now nicely animated, while the lagging behind
hover effet has been removed.

REVIEW: 126673