Fix double-click in Widget Explorer not running config initialization scripts.
This is better/simpler code, but doesn't fix that Plasma::Applet::loadPlasmoid()
instanciates applets but doesn't run scripts - which is difficult to achieve
because the scripting environment is designed around knowledge of a containment.
Perhaps scripts execution should be moved from ShellCorona to Applet and happen
the first time an applet is parented to a containment.

CCMAIL:notmart@gmail.com