[notifications] Place the popup directly when it is displayed
Now with the geometry being correct right after the popup is
being displayed, it can be placed directly to its exact intended
position (compared to sliding from default position because the
geometry was not known).

This should fix any issues left with the popups flying across
the screen and removes tons of code as a bonus.

REVIEW: 126668