Remove completely duplicated function
Remove almost completely duplicated function.
One was used for right clicking on expanded items, the other for full
compact items.

A missing action for expanded items is added so the two menus are now
consistent.

Global co-ordinates are used as args for menus everywhere so we can
share the mapping code SNIs currently use.

REVIEW: 125973