[Media Controller] Support multiple players
If more than one player is currently running, a ComboBox will allow to choose
the player to control. Doing so will make all controls (ie. the slider, album art,
player buttons as well as the tooltip and system tray icon, active state) reflect
the current player's state.

By default the "@multiplex" source is still used which tries to choose the player
automatically (I guess...) and it also reverts to that if the current player quits.