Reset seek slider position when changing songs
The attached patch retrieves the position from the player when:
- a new maximum is set to the slider (most cases)
- the the slider is already at the end (when the same song is played again)

Note that adding 1 sec to the slider when the remaining is less than a second
sets the slider to maximum. Handling at that moment seemed to be too soon when
I tried it. Waiting for the next iteration looked to be a better transition

REVIEW: 126779
BUG: 358135
FIXED-IN: 5.6.0