move all the lockscreen ui in a separate file
load the lockscreen wallpaper immediately and don't bother
to load the actual ui.
after it's done (we get a frameswapped) load the rest of
the ui in a loader.

this will end up in the lockscreen loading actually
slower in absolute times, but in the *perception* it
will look like radically faster.

REVIEW:126837