[Device Notifier] Provide inline feedback
Instead of showing a detached status bar with the device name, show the message
below the device in the list.

When removing a device, device notifier will pop up and a "You can now safely remove
this device" message will show up with a highlight gradually fading away until 5 seconds
later the device is removed from the list.

REVIEW: 126688