Implement rename() in a way that avoids losing custom icon positions.
If built against KIO v5.20+ we use a new version of the KDirNotify
file rename notification interface that allows providing enough
information to allow KCoreDirLister to announce a file change
instead of relisting the affected file(s) and causing position data
to be lost. We also make use of a new mechanism for informing KIO that
kio_desktop will in fact take care of producing the rename notification
by itself, avoiding KIO::SimpleJob doing it as well (and losing the old,
lossy interface since it doesn't have the same information we have
available in kio_desktop).

Reviewed-by: faure@kde.org
BUG:359596