[digital-clock] Fix display of seconds with certain locales
Currently, "Show seconds" has no effect with certain locales (C e.g.)
that already contain the seconds in the ShortFormat.
The reason is that the seconds are only added to the format string
that's used to display the time if they are not part of timeFormatString
already (which is not used at all for the display).

This patch fixes that check and adds them only depending on the
showSeconds config option.

REVIEW: 127623