[Shell Runner] Add "Run in Terminal" option
This allows to run the given command in a terminal window.

The runner used to have a "Run as other user" option but since it's widget-based
and nowadays we usually discourage this and it's an even more advanced option,
I'm not really into coming up with a solution for that.

Differential Revision: https://phabricator.kde.org/D1591