[PanelView] Fix auto hide
The newly introduced HiddenStatus which is used by eg. the desktop pager to completely
hide itself when there is only one virtual desktop, has a higher value than
RequiresAttentionStatus for ABI compatibility reasons, so auto-hide was never actually
enabled if you had this applet (part of the default setup) with just one virtual desktop.

BUG: 362105
FIXED-IN: 5.6.5

Differential Revision: https://phabricator.kde.org/D1683