Add a prop to toggle the special grouping behavior for tasks in demanding-attention state.
Summary:
This is needed for subsequent feature work which needs to be able
to turn this off.

This review also includes a number of bug fixes to the model:
- Fix an off-by-one when explicitly breaking a group apart.
- Fix rowCount() returning the value for a top-level item with
  the same row when called for a group member.
- Fix updating/invalidating persistent indices endRemoveRows()
  doesn't know how to handle.
- Update data for former parent item ahead of child-reparenting
  insert when breaking a group apart.

None of these fixes actually matter for the applet view, but
they obviously improve correctness, and will matter for later
work.

Reviewers: davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D1838