Write out QT_SCREEN_SCALE_FACTORS instead of deprecated QT_DEVICE_PIXEL_RATIO
Summary:
We use QT_SCREEN_SCALE_FACTORS as opposed to QT_SCALE_FACTOR as we need
to use one that will *NOT* scale logical DPI according to the scale.

Scaling the fonts makes sense if you don't also set a logical DPI
manually, but we need to do that for both PlasmaShell which does it's
own thing, and for KDE4/GTK2 applications.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D1849