Fix kicker launching
starting the desktop file from URL doesn't make too much sense as it
means we parse the mimeType beore then finding it's a desktop file.

Worse, this is broken where we only have a name, for example if you add
a KCM as a favourite.

It's done to avoid the blocking static methods, however there is an
alternate trick to save it blocking.

Reviewed-by: David Faure/Kai Uwe Broulik