Fix launchers getting lost on DND resorts when using 5.6-style config.
This is an unfortunate regression that crept in via 7eff45a98f4e5 the
just the other day: Legacy config can have query items other than
iconData in launcher URLs, which also need to be stripped before doing
the equivalence test matching up launcher-backed window tasks with
launcher URLs. Back to venerable launcherUrlsMatch() as originally --
the real speed optimization is in fetching the right role, anyway, and
as it turns out we really can't save that one call to adjusted() as it
was necessary.

BUG:364492