Invalidate cached launcher count also when the source model changes.
When a launcher was removed, the cache was invalidated at the time the
top-level proxy removes the row for it. However this happens in
response to rowsAboutToBeRemoved from the source model, at which time
launcherList() still contains the launcher at the row about to be
removed. While invalidating the cache in response to rows being removed
from the top-level proxy is correct (for the case where a window task
shadowing a launcher is removed), to handle a launcher removal we also
need to invalidate it when the launcher list changes.

CCBUG:364537