Do tests for startup notification filtering against state-filtered task list.
We don't show a startup notification if we already have a matching
window -- however, this test was done against all window tasks, even
if some of them are being filtered out (e.g. by virtual desktop).
Now "do we have a window?" is checked against the filtered window
tasks, so a startup notification for $app is shown even if $app
already has a window on a different desktop.