Don't set both font and pixel size in DigitalClock
Summary:
Digital Clock sets fontSizeMode, this scales the font to fit between a
tiny size and a the given size.

In older Qt this used to be between 0 and pixel size - but because
Plasma.Components.Label already set the pointSize we had to set both.

In Qt 5.6 it seems to handle either type of font size being set, so we can
keep with just one.

Test Plan:
Clock still scales to fill space
Clock still switches into compact mode if needed

Reviewers: #plasma, sebas

Reviewed By: #plasma, sebas

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D1974