Allow struts on panels between screen edges if WM is KWin
Summary:
KWin starting with 5.7 supports struts on panels between screen edges.
Thus we can start setting struts on such panels, it won't exclude a
complete screen. But we don't know how other window managers handle it
and it's in general a rather "dangerous" change.

Thus to not affect other window managers, we check whether KWin is
running and only allow struts on thus panels if KWin is running.
Unfortunately we need to test this every time we go into the code path
as the WM might have changed.

In case the user replaces the window manager at runtime this still can
result in a bad situation.

BUG: 94470
FIXED-IN: 5.8.0

Test Plan:
Tested whether it works in general in X11. Further testing
needed by X11, multi-screen users.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2164