Port Switch Window containment action to libtaskmanager-ng.
Summary:
Unfortunately this alone won't make it work on Wayland, as it's
still based around storing X11 window ids in QActions.

For it to work on Wayland, it should store model indices instead.
However that requires us to be able to refresh contextualActions()
while the menu is open, which the Plasma framework does not yet
allow.

We obviously need to address this later. For now, let's drop
the old backend and save some CPU and memory.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2165