ScriptEngine exports the list of activities known to Plasma
Summary:
There is a small time-window when Plasma's internal data
regarding activities is not synchronized with the internal
data of the KActivities library.

This leads to Plasma reporting a different list of activities
than the list of activities it has containments for.

This patch changes the activity list that the ScriptEngine reports
to be the one that Plasma knows about.

See: https://bugs.kde.org/show_bug.cgi?id=365606

Reviewers: #plasma, davidedmundson, sebas, mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2187