[ksplash] Better support for Wayland with multiple screens
Summary:
On Wayland ksplash only showed on one screen. This is due to QtWayland
not passing any hint to the compositor on which output to show the
fullscreen window. Thus the compositor placed all windows on the first
output.

This change uses KWayland to work around this problem by performing
a deeper integration. It doesn't use fullscreen windows any more, but
normal windows in combination with the PlasmaShell interface. That way
the windows can perform absolute positioning.

As the window is no longer fullscreen, it's marked as OnScreenDisplay
which makes sure that KWin raises it on top of all other windows. Also
ensures that opened windows aren't raised above it as used to happen.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2242