[Kicker] Disable PackageKit integration (aka "Uninstall app") context menu entry
Currently we cannot populate the context menu asynchronously which means we
block waiting for PackageKit which takes multiple seconds. While fixing the missing
include was correct, the result is a terrible regression in a bugfix release.

Once changes to the context menu can be notified by the Kicker backend, ie. 5.8,
this feature will return, sorry.

As discussed with Eike